Switch to melpaBuild for generated packages
trivialBuild results in files installed directly into site-lisp/,
without package metadata. This breaks ELPA packages whose dependencies
we've replaced: package.el refuses to activate the ELPA package because
as far as it knows its dependency is not installed.

melpaBuild seems the easiest way of generating an activatable package
that works with packages we leave untouched.