Use Doom profile loader to load our profile
This seems like a relatively sane way of loading our profile from the
Nix store. The alternatives I see are patching Doom (which I'm trying to
avoid) or setting DOOMLOCALDIR. If we set DOOMLOCALDIR, Doom sets enough
variables based on it it would be a pain to clean them up after we gain
control.

Using the profile loader lets us set doom-profile-data-dir early, which
currently only contains the profile and env file.