Disable doom-initialize-core-packages at runtime
...by moving the existing hack from the CLI to init.el.

It's not really necessary to do this as early as init.el, but I'm coming
around to using init.el modifications as an alternative to using Doom's
profile loader. So we might as well do this in the same place.

This is not sufficient to make doom/help-packages work, but gets us one
step closer...