Prune dependencies of example doomdir
This now matters for CI, as the example is still built using fetchGit
instead of the github fetcher.

The example was doing double duty as an example and a stress-test,
enabling some problematic or interesting modules. That is no longer
necessary (once the branch with dynamically generated init.el with all
modules/flags enabled lands).