CI: experimental flake-update workflow
I want to use a matrix strategy for the check and cachix workflows, so I
need to not just commit but also push the updated flake.lock somewhere
other than `main`, so multiple workers can check out the pending change.

Try the simplest possible thing first, expecting it to fail with
permission issues.