Expose experimentalFetchTree in home-manager
This should work around #14, which breaks fetches on newer versions of
Nix. I was hesitant to expose this as Nix changes may break things with
no warning, but it gives users already willing to run newer versions of
Nix a workaround for the fetch failures they're currently getting.

(And the feature itself seems to work fine on current Nix, since CI is
using it...)