Improve error messages for unfetchable packages
Split out the check for "pin" (which already had a dedicated check, but
it was mostly unreachable) from the check for "repo URL", and make both
error messages more useful.

Remove a TODO while I'm there, because it looks unachievable: we should
only hit these checks when using our generated derivation, which should
mean we're also generating `src`, which means we really do need a pin,
as Nix will otherwise refuse to fetch in pure evalution mode.

For #20