Never do shallow fetches
They seem to generally not work if the rev we're fetching is not
current. That failure mode is particularly problematic because already
having the rev cached avoids the failure: fetches work initially, when
the pinned commit is the current one, then break after upstream commits.

I could disable repository caching in CI, but that would turn it into a
game of whack-a-mole: this breaks on upstream repository commits, not
flake updates.

The downside is increased disk usage. This can be partially mitigated by
enabling experimentalFetchTree, which fetches snapshots.

Force shallow false to work around
https://github.com/NixOS/nix/issues/11012 (Nix 2.23 default to shallow
fetches). We should be able to remove that if Nix adds the fallback
they're planning to add.

Hopefully fixes #14 once and for all.