Drop magit pin again
Whoops: I assumed the melpa branch HEAD Doom was using was actually
working for Doom, but it stopped working with straight.el after that
branch was regenerated after an update to the main branch. Should have
tested / reported upstream.

Doom pinned git-commit to the main branch HEAD and magit to the melpa
branch HEAD~1, which results in the same commit, so I can drop this
again (and have to, to pull in the git-commit update...).