Pin forge
Relative to the version of Doom we currently pull in, this is an upgrade
from 0.4.3 to 0.4.4 (no significant changes).

Relative to the version Doom currently pulls in, this is a downgrade of
one commit, because it looks like that commit depends on a snapshot of
transient newer than what emacs-overlay currently provides.

https://github.com/marienz/nix-doom-emacs-unstraightened/issues/37
tracks removing this pin again.