Initialize profile in intended order
Doom loads the partial profile
snippets (90-loaddefs-unstraightened.auto.el and friends) in
lexicographical order, but calls their closures in
doom-profile-generators order. That means our additions load first.

This does not currently break anything I noticed, but it is not obvious
or intended: sort the list after we add to it to get the expected load
order.

(The next commit changes how we load init.el, and without this fix we
would load that later than intended: it still did not break anything I
noticed but it does get confusing...)