Support literate config / `doom +org tangle`
Adds `tangleArgs`. When set, runs `doom +org tangle` in a symlinked copy
of DOOMDIR, using that instead of the original for everything else.

I expect this to be usable as an alternative to `doom :config literate`,
which (because it does its work from `doom-before-sync-hook`) has no
effect.

We currently (somewhat hackishly) run `doom` with just an `init.el`
enabling `:lang org` (the bare minimum for Doom to provide the `+org
tangle` command), not a full profile. This means the user can tangle
their init.el, but also that Doom uses org-mode from upstream emacs, not
its pinned version. If this turns out to be a problem, we should be able
to build an actual Doom profile and `doom` wrapper first.