Fix tangleArgs from home-manager
toInit was using `lib` from the nixpkgs input. That makes sense when
running checks, but when I reused it for tangling I missed that that
input may not be available (the README recommends unsetting it as it
should not be used).

Fix it by using `lib` from the pkgs instance we're building from
instead.

Should fix #43.