Always override commit when pinned
For pinned packages with a melpaBuild from emacs-overlay, emacs-overlay
updates were resulting in derivations that differ only in their "commit"
attribute. That gets passed to melpa2nix, which puts it in the -pkg.el
file so it can be used by describe-package and friends.

So this makes those descriptions correct, but it should also cut down on
spurious rebuilds when emacs-overlay updates.