Improve dependency extraction
It turns out package-dir-info has some problems: the way it locates
-pkg.el files meant it was not actually doing so for us (it relied on
the directory name matching the package name), and it then falls back to
the first plausible-looking .el files: that may not be the right one if
we build multiple packages out of the same repository checkout.

This causes a build failure after the next emacs-overlay update, which
drops company-lean and lean-mode from Melpa (meaning our dependency
extraction needs to work on those).

Try to fix it by strictly extracting dependency info from the NAME.el
file.

Make dependency extraction failure fail the build. We run it on few
enough packages I think this is worth it (to catch bugs).

This has two annoyances:

- It uses some package.el internals that moved in Emacs 30, so we need
  two codepaths.
- We have one package (x-face-e21) for which the NAME.el file is not in
  the root. For now, add a hack that looks in one level of subdir.
  The proper fix would be to use :files from the recipe...

This is not sufficient to make lean-mode build, but should be closer.