Add custom recipe :files for lean-mode
Melpa (and therefore emacs-overlay) is dropping lean-mode and
company-lean. We really do need the :files bit of the dropped recipe,
not just the url: these two packages build from the same repository.
`nix flake check` does not notice the duplication, but a full (cachix)
build fails because we try to compile company-lean.el when building
lean-mode, which does not depend on company.

Add a bit of plumbing to pass :files into the recipe passed to
melpaBuild, which seems to be sufficient to fix the build failure.