Improve DOOMPROFILELOADFILE handling
Use the path without .el/.elc suffix. Emacs passes this to `load`, which
tries with .elc first, then .el, so this does what we want. And this is
both closer to what Doom normally does and to what `load` seems to
expect.

When writing, explicitly use the path with .el. Previously, we ended up
writing the uncompiled loader to the .elc, and then overwriting it with
the byte-compiled version. Having the original around is convenient when
debugging.