Make extraBinPackages available to just Emacs
When I added extraBinPackages, I missed that the emacsWithPackages
machinery in nixpkgs already supports something similar: any extra
packages get their bin/ directory added to the one in
emacs-packages-deps, which it adds to exec-path.

This is subtly different from adding them to PATH: Emacs itself (from
functions like `executable-find`) respects `exec-path`, but child
processes invoked via a shell do not.

I think that is actually closer to what I want for extraBinPackages. I
tried to check the use cases mentioned in the issue tracker and it looks
like they also have Emacs invoke processes directly (without a shell in
the middle).

But admittedly I'm also changing this because the implementation is
simpler.

This arguably makes extraBinPackages unnecessary, since extraPackages
provides a superset of its functionality. I'm keeping extraBinPackages
to not break existing configs, and because in a home-manager
configuration the separate list of packages (instead of function) seems
nicer.