Fix CI failure for lsp-haskell
Upstream dropped a dependency on haskell-mode still required by the
version we are pinned to.

For now, fix this by forcing the dependency back. But it would be nicer
to be able to use our dependency-extracting logic, so we lose the
dependency immediately once Doom catches up.

I would still want to remove the dependency-extraction once no longer
needed, as it is slow... so maybe this isn't too bad.