Pin package-lint
See https://github.com/purcell/package-lint/issues/285.

Doom added a pin to work around the same problem, but we hit an edge
case where it does not apply:

When building our example init.el, which has `:lang emacs-lisp` and does
not have `:checkers syntax`. `:lang emacs-lisp` pulls in
flycheck-package when `(modulep! :checkers syntax -flymake)`, which
returns `t` while `(modulep! :checkers syntax)` returns `nil`: this
seems to contradict that function's documentation (I'll ask there if the
bug is in the documentation or the implementation).

Doom's pin is set in the `:checkers syntax` module, so with that module
disabled it does not apply.

Unconditionally pin to the same revision Doom pinned to.