Write non-nativecomp builds to cachix
The "magic" nix cache is about to go away. I will try to use cachix as a
replacement.

To allow `check` to still run with some caching on forks, I will try to
set up cachix in read-only mode for that workflow (the same setup as any
other public cache), while the CI / cachix workflows use the cachix auth
secret and push to the cache. This is somewhat redundant/inefficient but
seems like it should work (and should make the `check` workflow work the
same for me as it does for contributors).

Implement the "write to cachix" half of that first, so I can confirm the
cache gets used when I switch `check` over.