Always specify a ref to work around caching issue
This reapplies 2f89220bdcbf8118db82753672684d8e27d4fc4c, which was
reverted in 9205f2878ff2d173b91cf27f594d4f9a983b639f.

This works around https://github.com/NixOS/nix/issues/10773: a Nix cache
that Nix does not update after it becomes invalid.

We reverted this because of #14, but this should not have contributed:
that issue was mitigated by not doing shallow fetches.

I did not notice I had not reapplied this because I had bumped
tarball-ttl locally, which also works around it. But I had bumped it to
1 year, which means I recently started to become affected again...