Fetch ws-butler from original source for now
emacs-overlay is pulling in a change in source for this, which is
failing to fetch (and in parallel Doom is updating to a newer revision,
but at first glance that new revision should be in both repos).

Fall back to the old location. This may break again if Doom follows the
move but we can deal with that if/when it happens.