Fix build of wanderlust with new melpa-build
nixpkgs pulled in a newer version of melpa-build, which seems to newly
enforce that packages contain a library with the same name as the
package. Wanderlust itself and three related libraries violate this.

Pull in patches that fix this. As far as I know this has no noticeable
effect on older melpa-build.