Don't repeatedly recompile the profile loader
Unstraightened always triggers async native compilation of the profile
loader on startup: Emacs looks for a native-compiled file before Doom
has updated native-comp-eln-load-path, but async native compilation uses
Doom's updated value, so the native-compiled file ends up in a location
where Emacs will never find it again.

If I understand correctly, this requires a function in the profile
loader to trigger: that means any more reasonable user of Doom's profile
loader will not trigger this, which makes me hesitant to try to fix this
in Doom's profile loader (backported from Doom 3.0...).

The easiest fix on our end would be "remove the byte-compiled profile
loader", but I prefer not to deviate from Doom if not strictly
necessary.

This implements what I think is the second-easiest fix: disable
native-comp for the profile loader.

The way it does it is not the prettiest, but I expect to revisit this
after Doom 3.0: either because this will need to be rewritten, because
it's unnecessary, or because it can more reasonably be upstreamed.