Skip testing :tools upload
It pulls in "ssh-deploy", whose upstream seems to have disappeared.

The module has no other effect / does not configure additional packages.
So this seems the easiest way to unblock CI, and we don't lose coverage
of anything else.

Do this "properly" by adding filtering to the full-init doomscript
instead of doing more post-processing on the Nix side: post-processing
would have worked (this module has no flags) but is getting out of hand.

For #66.