Pass necessary --init-directory when restarting
doom/restart-and-restore (and likely other Doom-restarting functions
typically bound to `SPC q` was restarting into a vanilla Emacs.

The problem is that it restarts the emacsWithPackages-provided "emacs"
binary instead of our wrapper, which would add the `--init-directory` we
need to initialize Doom. That wrapper also sets a few environment
variables, but those seem to be making it through unscathed.

It looks like that happens because emacsWithPackages sets
invocation-path and invocation-name (as of
https://github.com/NixOS/nixpkgs/commit/5f9cbf9898b86c8151dde657c2023f3afdb714bf,
before that this might not have worked either but would have failed
differently).

Let's fix this for now with restart-emacs advice that adds the necessary
argument.

Although I think we cannot sensibly change what emacsWithPackages sets
invocation-path / invocation-name to, another approach (which did not
occur to me until after I had gotten this one working...) would be to
set invocation-path / invocation name ourselves (from init.el, after
emacsWithPackages has set them). We can try that if this approach
breaks.

Should fix #86.