enable LSP_USE_PLISTS (with an option to opt out)
Defaults to enabled in Doom, so enable by default.

Try to only patch lsp-mode-related packages to avoid rebuilding the
world (and having to build everything in both modes in CI).

Provide an opt-out: both because this will break things if I missed
something in the "detect which packages to patch" logic, and because
even though lsp-mode upstream seems to recommend this as being more
performant, it is not their default (and I'm not sure if their only
reason for that is the bytecode compatibility breakage).

Fixes #83