Accept (and ignore) `:env` on `package!`
Doom describes this as a "temporary measure", used for LSP_USE_PLISTS.

I cannot easily handle this the way Doom does, and have a Nix-side
solution for LSP_USE_PLISTS mostly done. For now, assume this is indeed
both temporary and unlikely to spread to other packages (and variables),
meaning we should be safe ignoring it.