Apply elisp-packages as separate overlay steps
This is a rather large change (and it looks even larger because
elisp-packages-late.nix does not show as the rename of
elisp-packages.nix it largely is because of some whitespace changes) and
it's not quite diff-clean but I think it's worth it:

- It turns the three-way selection of "which package to use" (original,
  custom, or generated) into a two-way selection
- It lets elisp-packages further override those generated packages
- It prepares for adding some more overrideScope calls
  (for a user-provided overlay and lsp-use-plists)

It's a large change because neither elisp-packages-early nor
elisp-packages-late is equivalent to the previous elisp-packages:

- elisp-packages-early runs unconditionally (overrides here cannot tell
  if they will be pinned). This is generally not desirable, as our
  packages may be pulled in indirectly without Doom pinning them.
  Only one override made it here: a package rename.

- elisp-packages-late runs after pins, so its overrides must now
  inherit (at least) src from the original.

So I cleaned up the existing overrides (frequently converting from a
trivial melpaBuild to overrideAttrs of a generated package) and moved
them to -late.

There were some small changes to derivations, which generally stem from
those cleanups: we now more consistently use auto-detected dependencies,
for example.

(There may be one or two more melpaBuild instances in nix-packages-late
that can be converted to overrideAttrs, that can wait for later, what's
there works.)