Fail if doomDir is not readable
doomDir can be pointed outside the Nix store: this does not
automatically fail the build, instead we build as if we were handed an
empty config.

(Nix's sandbox presumably runs us in a cleaned-up mount namespace, and
Doom is generally tolerant of missing files in DOOMDIR...)

We should probably do some type-checking on the Nix side, but let's make
sure we fail noisily if DOOMDIR is missing outright.

(We could check for init.el or similar, but pointing us at a non-empty
directory not containing that seems harder to do accidentally. This
check seems sufficient.)

For #94.