Require that doomDir is in store
`path` means "absolute path", but it can be an absolute path outside the
store. We use this at build time so we need it to be in store.

If this breaks your configuration: please file an issue! This is
intended to avoid more confusing failures (particularly silently
ignoring the configuration if it is not visible at build time): it is
not intended to break anything that currently works.

For #94.