Handle preBuild being null
A handful of packages in nixpkgs set preBuild conditionally (based on
version) but set it to null when they intend to leave it unchanged (and
the original they patch do not have the attribute). That breaks our
further patching.

Work around this by checking for null.

I don't do this more broadly (for the overrides in
elisp-packages-late.nix) because I do not think we should need to be
this careful. We can add more workarounds if we find module combinations
that actually trigger the problem there.